Implemented signal for getting Fourier bases

+++ examples/plot/src/main.rs
@@ -10,7 +10,7 @@ fn main() {
     let xs = sample(&step(), -10.0, 10.0, 0.2);
 
     let idx: Vec<usize> = (0..xs.len()).collect();
-    let ys: Vec<f32> = xs.iter().map(|x| x.re).collect();
+    let ys: Vec<f64> = xs.iter().map(|x| x.re).collect();
     let mut fg = Figure::new();
     fg.axes2d().lines(&idx, &ys, &[Color("red")]);
     fg.show();
